pyproject.toml
[project]
name = "tools"
version = "0.1"
requires-python = ">=3.10"
dependencies = [
    "pytest-playwright",
    "pytest-unused-port",
    "markdown",
]
